Update a few things;
Using Babel now instead of other JS syntax def. I'm probably going to
keep updating which syntax definition I use, but Babel seems like it's
playing for keeps. I added PHPlint but disabled it. When I learn more
of how to use it, I will enable it and make it a bit friendlier. I
disabled CSSLint because I haven't taken the time to make a csslintrc
file which works to my liking.

Updated to Afterglow-magenta! I love this theme, and it really blends well with Hopscotch.

Added an INI and a WordPressDev package